+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/group/detail/GroupDetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.happy.friendogly.domain.mapper.toPresentation
+import com.happy.friendogly.presentation.ui.group.mapper.toPresentation
 import com.happy.friendogly.domain.usecase.GetClubUseCase
 import com.happy.friendogly.domain.usecase.PostClubMemberUseCase
 import com.happy.friendogly.presentation.base.BaseViewModel
