+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/group/list/GroupListViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.happy.friendogly.domain.mapper.toDomain
-import com.happy.friendogly.domain.mapper.toGenders
-import com.happy.friendogly.domain.mapper.toPresentation
-import com.happy.friendogly.domain.mapper.toSizeTypes
+import com.happy.friendogly.presentation.ui.group.mapper.toDomain
+import com.happy.friendogly.presentation.ui.group.mapper.toGenders
+import com.happy.friendogly.presentation.ui.group.mapper.toPresentation
+import com.happy.friendogly.presentation.ui.group.mapper.toSizeTypes
 import com.happy.friendogly.domain.model.UserAddress
 import com.happy.friendogly.domain.usecase.GetAddressUseCase
 import com.happy.friendogly.domain.usecase.GetSearchingClubsUseCase
