+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/group/mapper/ClubFilterMapper.kt
@@ -1,4 +1,4 @@
-package com.happy.friendogly.domain.mapper
+package com.happy.friendogly.presentation.ui.group.mapper
 
 import com.happy.friendogly.domain.model.ClubFilterCondition
 import com.happy.friendogly.domain.model.Gender
