+++ android/app/src/main/java/com/happy/friendogly/presentation/ui/group/mapper/ClubMemberMapper.kt
@@ -1,4 +1,4 @@
-package com.happy.friendogly.domain.mapper
+package com.happy.friendogly.presentation.ui.group.mapper
 
 import com.happy.friendogly.domain.model.ClubMember
 import com.happy.friendogly.domain.model.ClubPet
